mips: a preliminary porting to the loongson2f yl8089 board
which is based on the work of yanhua<yanh@lemote.com>,
but the original porting is out-of-date.

Todo list:
        1. (known problem) The SM712 graphics do output but not prefectly.
        2. Reorganize the code.

A straight-forward guide to use U-Boot on yl8089

Build the U-Boot binary image:
1. (experimential, dangerous) If you want to program it into flashrom, change
   CONFIG_SYS_TEXT_BASE to 0xBFC00000(in board/lemote/yl8089/config.mk)
2. 'make menuconfig'
3. 'make CROSS_COMPILE=mipsel-linux-'
This should generate u-boot(.bin), the u-boot can be loaded by pmon and the
u-boot.bin can be programed into bios eeprom.

How to load kernel:
The u-boot can support elf and binary file load
1. To load from IDE/USB ext2 partitions
   #ext2load ide/usb 0 [80200000] /boot/vmlinux.bin
   #go [boot entry] [kernel args]
2. To load from tftp server
   First you need to config ip address, either manual set or by dhcp
   a)DHCP
   #dhcp
       This should set 'serverip' 'ipaddr' 'gatewayip' and .etc.
   b)manual set ip address
   #set serverip  172.16.0.30
   #set ipaddr    172.16.0.30
   #set gatewayip 172.16.1.254
   Once you have done the ip setting, you can do the actual loading
   #tftp 80200000 vmlinux.bin
   #go [boot entry] [kernel args]

USB-supporting
The usb support is limited. Now only the NEC host is supported(Later may
all usb ports be enabled)
The usb is not configed by default. Below are a brief description for
usb.
1. usb start
   ## this will initialize the usb system and scan the bus for connected
   devices such as usb kbd and usb storages
2. usb info
   ## this will show usb information
3. usb storage
   ## show usb disk capacity and other info if there is one
after usb start, if you want to usb kbd, exec the following command
   #set stdin vga,serial,usbkbd

BIOS(u-boot) Updating
There is environment script can help to do it
1. prepare the image
   Assume the image is named as u-boot.bin and is stored in
   /tftpboot/u-boot.bin
2. run ub
   ##This will load u-boot.bin from tftp and erase/program the eeprom.
   for other updating method try the 'help' command.